evaluation/requirements.txt
# Core Dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=9.0.0
tqdm>=4.64.0
opencv-python>=4.6.0

# Model Frameworks
ultralytics>=8.0.0              # YOLO v8
segmentation-models-pytorch>=0.3.0  # U-Net, DeepLabV3+

# Optional: Detectron2 for Mask R-CNN (install separately)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Data Processing
scikit-image>=0.19.0
albumentations>=1.3.0
matplotlib>=3.5.0

# Utilities
pyyaml>=6.0
tensorboard>=2.10.0            # For training visualization (optional)

# Note: For Windows users, set num_workers=0 in config.py
# Note: For Detectron2, install separately following their official instructions
